refactor: extract consent hook
